[Profiler] Bug fixes (#2309)
This PR contains refactoring and bug fixes for `create-visuals.py`, and
an updated README with instructions for using the profiler. More
specifically, `create-visuals.py` used to not account for "cells" as
distinct elements of the stack (so, the flame graph would not reflect it
when a non-main cell was taking extra cycles between groups), but this
PR fixes that issue.