[Profiler] Pass to produce instrumentation code (#2310)
This PR contains (in order of importance):

- `instrument.rs`: a new pass (`-p instrument`) that adds
instrumentation to each group.
- A new primitive (`std_inst_wire`) which functionally works the same as
`std_wire`, but specifically is *not* a combinational component. This is
important for preventing optimizations from removing any cells of said
primitive.
- `compare-instrumentation.sh`: a script to make sure that
instrumentation did not affect latency.
- Changes to test expect files to include the new `std_inst_wire`
primitive.

## Instrumentation
The `instrument` pass iterates over all groups. For a group `foo`, it
produces (1) `foo_inst`, a `std_inst_wire` cell, and (2) an assignment
within `foo` that sets `foo_inst` to 1. This makes `foo_inst` have 1
whenever the group `foo` is active.