[Cider] Correct `@control`/`@data` processing and add some optional logging (#2319)
Sorry for the larger PR here, a bunch of stuff ended up getting all
tangled together.

This PR adds:
- a check for undefined guards after convergence. This is currently
disabled by default as the changes to `@control` mean that this will
basically never be relevant for normal programs.
- redoes internal error handling to make proper error messages more
enforceable by the type system
- Adds a configuration struct for runtime checks/behaviors
- Adds optional debug logging via `--debug-logging` which prints out the
assignments that fire and the results they propagate alongside the
implicit zero assignments. Might be useful for some people under some
circumstances?
- Adjust the `@control` port process to match that used by the compiler.
Consequently for most programs we simulate, every port is `@control`.
This means some error cases are no longer errors and also reintroduces
transient conflicts on some programs.
- As a spot fix to the transient conflicts, when detected we will now
check if the guard for the first assignment still evaluates to true and
if not we take the later assignment and continue onward, otherwise an
error is raised. In all likelihood this won't mitigate every instance of
the transient conflict problem but seems to address it for our current
programs.