Write docs for `calyx-pass-explorer` (#2341)
This PR adds a documentation page for the `calyx-pass-explorer` tool,
serving as a gentle introduction to its usage as well as an overview of
its code.

- I've added `mdbook-callouts` as a way to get Obsidian-style quote
blocks. Specifically, I'm using this for the `[!TIP]` environment. I can
remove this if we don't want another dependency.
- I did some unrelated things (to `docs/contributors.md`,
`docs/github.md`, and `docs/intro.md`) that I can move into another PR
if that would be better.