[Profiler] Profiling support for parallel programs (#2370)
This PR contains:
- The `profiler-instrumentation` pass is now updated to contain four
types of probes in order to capture call site information.
  - `group_probe`: A probe that is high when a group is active.
- `primitive_probe`: A probe that is high when a primitive cell is
active.
- `se_probe`: A probe that is high when a group that structurally
enables another group is active.
- `cell_probe`: A probe that is high when a group that invokes a
non-primitive cell is active.
- `profiler-process.py`: an updated VCD postprocessing script that
generates `.folded` and `.dot` files:
- A "flattened" flame graph, where each parallel arm gets its own
"cycle". So, if arm A and arm B were executing on a single cycle, the
flame graph would account for a cycle in arm A and a cycle in arm B.
- A scaled flame graph, where a cycle is divided between the parallel
arms in execution. So, if arm A and arm B were executing on a single
cycle, the flame graph would account for 0.5 cycles in arm A and 0.5
cycles in arm B.
- `aggregate.dot`: A tree summary of the execution of the program. Nodes
(groups and cells) are labeled with the number of times the node was a
leaf, and edges are labeled with the number of cycles that edge was
activated.
- `rank{i}.dot`: A tree representation of the `i`th most active stack
picture. `rankings.csv` lists the specific cycles that each ranked tree
was active for.
- Updated fud2 support (`fud2/scripts/profiler.rhai`)
- Utility scripts to help convert generated `.folded` and `.dot` files
into visualizations (`svg` and `png` files respectively)
- Updated tests/cleanup of old tests that relied on scripts that no
longer exist

As usual, I'd really appreciate any suggestions or thoughts!!