[Cider] Combinational Component support (#2372)
Small PR which adds support for combinational components and makes some
minor adjustments in a few other places.

The notable difference for the convergence algorithm is that we now sort
the program counter by containment, before iterating over it. This means
we will visit parents before children during convergence and is
necessary for race detection in the presence of invoke chains. While it
is unnecessary when not doing race detection, I've currently elected to
do it unconditionally for simplicity and consistency's sake, but if it
becomes necessary we can move it behind the race detection flag.