Generate queue `.data` files according to custom command size (#2375)
Small edits to `gen_test_data.sh` to generate `.data` files with
argument-specified lengths of commands, and a specified output
directory. If no arguments are provided, the size and the output
directory are set to the default (20000 and `<SCRIPT_DIR>/../tests`).
It's a bit hacky, but since I only need it for profiler debugging
purposes, I hope it's ok?