[Metadata] Initial redesign (#2396)
An initial redesign of the metadata table. This PR ended up a little
chunky, so to break things down it does the following:
- Introduces a `SetAttribute` to Calyx which stores multiple unique
numbers rather than only one. The syntax is `@Attribute_name{ n1, n2, n3
}` and `"attribute_name"={ n1, n2, n3 }` depending on the context.
- Move `pos` from `Attribute` to `SetAttribute`. If a `pos` attribute is
written with the standard syntax, the compiler will emit a warning and
ignore the attribute. E.g. `@pos(1)` is incorrect and will be ignored
since it should be `@pos{1}`.
- Also adds an `Unknown` set attribute, which works the same as its
corresponding variant in the standard attributes
- Adds a new `MetadataTable` construct which is parsed in a `fileinfo #{
.. }#` block and made available in the `Workspace` and `Context`
structures as an optional field.
- At present, the table maps file identifiers to paths and positions to
pairs of files and line numbers. We will likely want to extend this in
the future.
- The actual syntax of the table is a very basic placeholder which we
can iterate on later as needed
- There is currently a limited metadata lookup construct which will read
the file contents into memory and return the given lines. It's pretty
brittle right now
- We may in the future want a special signifier for the current file /
input stream as calyx->calyx mappings will want this. Though perhaps
there is a better way of approaching this
- Also moves the cider index implementation into a separate crate
`cider-idx` as we were initially going to use that with the new table,
but we moved away from that approach for the moment, so these changes
are largely irrelevant. Didn't want to do git magic since it was already
pushed and the branch is being used by @ayakayorihiro.