[Profiler] Fix compiler pass order for profiling (#2423)
The pass order used for profiling (both in the `-p profiler` pass and
set in fud2) was buggy since validation of the original program would
happen _after_ the profiler-specific compiler passes. When profiling, we
compile static groups before instrumentation (because we can't directly
instrument static groups), so validation fails when an (originally
static) group uses a static cell, such as `pipelined_mult`.

Thanks @pedropontesgarcia for trying out the profiler and letting me
know about this!