[Cider] Extend `step-over` for easier debugging of infinite/long groups (#2431)
Inspired by some attempts to debug an infinite loop or group running
longer than expected, this PR adds an optional bound `n` to the
`step-over` command. When present, the debugger will return control to
the user after `n` cycles if the group is still running. If not
provided, the behavior is the same as before, where it will continue to
run until the group exits. In the case of non-terminating groups, this
leads to an indefinite hang.