[Cider] Reimplement the display command (#2435)
Small PR which reimplements the display command which was left behind
during the overhaul. Rather than dump the ports for all cells in the
main components, as was the previous behavior, I've opted to make it
only dump the ports for cells which are in currently executing groups. A
limitation of this behavior is that groups which are on their last
cycle, i.e. their done ports are high, will not be counted as active.

Also, I decided to hash cons the guards finally, though I am not sure
why I decided to do this now.