Compiler pass to simplify comb group associated with a single invoke (#2429)
Edit: After 3/17's Calyx meeting, we decided to generalize the pass and
call it `constant-port-prop` instead. The idea is that we can expand
this pass to propagate constants instead of ports wherever we know it is
safe to do so.

The pass is updated to also replace uses of the cell's `go` ports with
`1'd1`.

---Original PR comment---

This PR contains a new pass `simplify-invoke-with` that finds all
redundant uses of cell `done` ports within a combinational group that is
associated with a single `invoke`, and replaces them with `1'd0`.

# Example
```
wires {
     comb group comb_group {
      wire.in = !invoked_cell.done ? 1'd1;
    }
}
control {
    invoke invoked_cell[]()() with comb_group;
}
```
In the above code snippet, `comb_group` is using `invoked_cell.done`
even though it doesn't have to, since the only usage of `comb_group` in
control is with the invocation of `invoked_cell`, and we know that
`invoked_cell.done` is false (`1'd0`) during the invocation of
`invoked_cell`. So, we can replace `invoked_cell.done` (within
`comb_group`) with `1'd0`.

Please let me know if anyone has any thoughts or suggestions about the
pass!