[Profiler] Improve control visualizations for timeline view, add TDCC-generated par group support, statistics on useful cycles (#2442)
This PR contains a lot of profiler visualization updates:
- Timeline view: 
- Removed FSM tracks (i.e. tracks to show the value of each FSM in the
cell), since I've gotten feedback that it was confusing since the user
doesn't know FSM values.
- Added a new "control" track that tracks updates to control registers.
Control registers are TDCC-generated registers, specifically any FSM
registers to track sequential control and `pd` registers to track par
arm termination.
- Flame graphs:
- `overview.svg` is a high-level "flamegraph" view categorizing each
cycle in the simulation as one of the following:
- `group/primitive`: at least one group or primitive is active, implying
the presence of "useful" work. Note that control registers *could* be
updated in cycles in this category, but have to be in parallel with a
group/primitive.
- `FSM`: the only "active" element is an FSM register update (could
involve multiple FSMs being updated)
- `par-done`: the only "active" element is an update to a `pd` register
- `mult-control`: both an FSM and a `pd` register are being updated.
I've never seen this in my example cases, but leaving this open as a
category because I think it's possible!
- `other`: This category is attributed to etc. It used to contain comb
groups, before I added support for them in #2440 . Now I see 0% `other`
but I'll keep the category around for at least a bit longer.
- The standard flame graphs now also include `par` groups so we can get
a sense of how much control registers within pars affect cycle counts.
- Cell statistics (`cell-stats.csv`)
- This CSV file now contains a percentage breakdown of the five
categories described above for each cell that was active. It also lists
the number of "useful" cycles (i.e. `group/primitive`) and the total
number of cycles.
- The CSV also now contains a "TOTAL" row aggregating information across
all cells.

## Noteworthy code changes
The changes to TDCC generate information about each `par` group,
specifically the name of the par group, and the group and `pd` register
names for each of the par group's arms.

## Next
I'm planning to do a reorganizing/refactoring of `profiler-process.py`
since it's become too big (1.6k lines of Python in one script... :sweat:
). I'll also do some case studies by hand and see what more
information/visualizations we need to provide via the profiler.