[fud2] Hard exit on emitter errors (#2447)
A quick stab at fixing #2348. 

It's a bit of a hack job, but I was trying to keep the code changes
minimal. The better solution would be returning emitter errors directly
rather than wrapping them in the `rhai::EvalAltResult::SystemError(..)`
enum.

End result is that if the emitter runs into an error, the emit process
will halt and return an error code rather than noting the error in the
ninja output.

edit: on further inspection, it seems like errors don't even always end
up in the ninja output under the old approach, so probably important to
get this fixed