Oracles compiler (#2460)
This PR closes #2458 by implementing a compiler that takes a JSON
representation of a Rio program and generates an oracle for it. To run
the bash file that generates expect files using the oracle compiler,
```
cd frontends/queues/test_data_gen
./gen_oracles.sh
```
We can run these 5 tests against the Calyx using runt to verify that
they pass, and so are indeed equivalent to the original oracles.