[Cider] Extensions to breakpoints (#2466)
This PR pulls @eliascxstro's work from this past semester. This generalizes our notion of breakpoints to support arbitrary control nodes instead of just group enables. Currently this works for breakpoints, but has not yet been ported to watchpoints, which remain limited to group enables.

There is still work to be done on improving the UI and making the path syntax understandable, but that can come at a later date.

---------

Co-authored-by: ec798 <ec798@cornell.edu>