[Profiler] Producing group statistics (#2472)
The VCD-processing scripts now produce statistics about groups,
including the min/max/average number of cycles the group was active for
on each enable. If `min` == `max`, then the group is a candidate to be
converted to `static`.

I also brought back changes to TDCC that output par parent-child info
that was removed.