[Cider] Add parent info to the `ControlNode`s (#2477)
Small set of changes I've been meaning to make for a while which
introduce a small redundant index to the `ControlNode` definition which
points to the node's parent. This allows for easier modification of the
program counter when advancing the execution state. As we previously had
to do a top down search from the root of the control tree.

I don't expect this to have much impact on performance as this only
executes during cycle advancement, but it wouldn't surprise me if it
marginally improved things since it removes the need for allocations
when determining the next node and bypasses searching the tree. All that
said, the primary motivation is just to clean up the code and make
things a little easier to deal with in the long run.

There's a bit more work needed to fully remove the old `SearchPath` code
but simply having backpointers should greatly simplify construction,
should the structure itself be useful.