fud2: Suppress SIGINT during Ninja execution
Fixes #2154. The idea is to suppress the default behavior of SIGINT *for
the fud2 process only*. SIGINT still goes to the child process, i.e., to
Ninja. This way, when you type ^C when running fud2, it kills the
underlying Ninja process as you expect, but fud2 still gets a chance to
clean up afterward---notably, by deleting the `.fud2` working directory.