fud2: Suppress SIGINT during Ninja execution (#2488)
Fixes #2154, which is about leaving the `.fud2` directory behind when
you interrupt execution with ^C.

The idea is to suppress the default behavior of SIGINT *for the fud2
process only*. SIGINT still goes to the child process, i.e., to Ninja.
This way, when you type ^C when running fud2, it kills the underlying
Ninja process as you expect, but fud2 still gets a chance to clean up
afterward---notably, by deleting the `.fud2` working directory.

Interestingly, the solution here is to register a no-op handler for
SIGINT. This is distinct from using SIG_IGN, which ignores the signal
for the entire process group (not what we want: it would make it
impossible to stop Ninja too!).

This seems to work here; interrupting with ^C looks like this:

```
$ fud2 --from ntt tests/correctness/ntt-pipeline/ntt-16.txt --to dat -s sim.data=tests/correctness/ntt-pipeline/ntt-16.txt.data --through verilator
^Cninja: build stopped: interrupted by user.
Error: ninja exited with exit status: 2
```

And the `.fud2` directory is indeed deleted. (We could consider making
the latter error message friendlier if we want.)

I would really appreciate it if someone else were to try this out
locally! And to check my thinking on the weirdo RAII-only struct for
managing the signal state.