[Fud 2] Report op splitting (#2496)
This PR atomizes the report extraction and parsing a little more by
splitting the compilation to synth-ready SV from the operation of
running Vivado synthesis on it and extracting reports.

This is the new `fud2` graph:

![image](https://github.com/user-attachments/assets/12afee03-c009-4649-829f-cd9577f0708a)

Further splitting would be desired, as proposed in this rough sketch:

![IMG_5349](https://github.com/user-attachments/assets/e21599ce-6db3-46dd-bddd-6192bc925f65)

Unfortunately, `fud2` is reliant on single files/text streams as inputs
and outputs of ops, and currently there is no ideal way to have a folder
(such as the Vivado output folder) as an input/output.

---------

Co-authored-by: Pedro Pontes García <pp457@havarti.cs.cornell.edu>