First cut at switching docs to recommend fud2
Just an attempt to convert all the low-hanging fruit I could find in the
docs to refer to fud2 instead of fud. I haven't done anything about the
actual fud docs themselves, the "experimental" label, or anything like
that.

Most of the changes are pretty mechanical. The only substantial bit of
new writing is in the "getting started" guide, where I added a
compressed overview of how to install fud2 (which is of course already
explained in more detail on the main fud2 docs page).