[Cider] Move the memory & register state into environment  (#2507)
First pass at moving the values of registers and memories into a big
vector, as with the other major constructs. This is a precursor needed
for the "entangled" memories hack that some of our examples want.

A lot of this PR is pretty mechanical and uninteresting. I did, however,
add a new map type to the `cider-idx` library which is meant to act
similar to `SparseSecondaryMap` but with dense storage. The primary use
case is when there are secondary key values that are mostly allocated in
contiguous blocks but potentially with gaps in between. In this case, I
am using this to store the clock ids for a given memory location, which
are mostly allocated in big blocks for memories. And sometimes allocated
in blocks for registers, depending on the order of register definitions
and constants (since constants lack clock ids).

There have also been some rewrites to the primitive trait with regard to
serializing memory/register data