[Profiler] TDCC utilization/hierarchy (#2510)
This PR incorporates initial work to add TDCC and PAR groups to
hierarchical visualization with the profiler. It adds a `fud2` script to
obtain a JSON hierarchy report, which can be then used by the profiler
(also with `fud2` through the options `synth-verilog.hier`, an absolute
path to a hierarchy JSON, and `synth-verilog.var`, one of `ff`, `lut`,
`llut`, `lutram`). The profiler is able to identify _some_ information
about TDCC and PAR groups, but is currently bad at identifying:
- Cell sharing with primitives
- Static groups
- The `multpipe` primitive

There is an open problem of how to tell the user that their groups or
primitives are being abstracted away or otherwise shared and repurposed.
More work and some creativity is needed here.

THINGS TO RUN:
```bash
$ fud2 <calyx-file>.futil --to json-report -o <synth-hierarchy>.json --through synth-verilog-to-hierarchy-json
$ fud2 <calyx-file>.futil -o <flamegraph>.svg --through profiler-area-util -s sim.data=/absolute/path/to/<sim-data>.data -s synth-verilog.hier=/absolute/path/to/<synth-hierarchy>.json -s synth-verilog.var=<one of ff, lut, llut, lutram> --dir <out-dir>
```

---------

Co-authored-by: Pedro Pontes García <pp457@havarti.cs.cornell.edu>