[Profiler] Pass to make enables unique, compute par threads, output path descriptors + correcting timeline view wrt threads (#2501)
This PR fixes the issue where the profiler's timeline view randomly
assigns threads to group enables active at the same time. It largely
makes two additions:

1. A new pass, `uniquefy-enables` (name suggestions are very welcome)
which creates a new group for each enable, so that we can differentiate
between two distinct enables of the same group. It also has options to
produce extra info files that assigns a thread ID to each enable group
and computes the path descriptor of the enable group as discussed in
#2273. (More writing on this pass below)
2. The profiler now reads in the outputted `par-thread-json` file and
uses the assigned thread id for each enable in the timeline view.

The `uniquefy-enables` pass is run in the profiler workflow (both in the
Calyx compiler `-p profiler` pass alias and in the `profiler` fud2
workflow) immediately before `profiler-instrumentation`. So, it assumes
that all `invoke`s are compiled into groups.

As always, any feedback is greatly appreciated!

## `uniquefy-enables`

This pass creates a unique version of each enable in the program. For
each enable in the program, the pass creates a new copy of the original
group with a fresh name by leveraging `builder.add_group()` (which
ensures that the group created does not have an overlapping name with
any existing groups). The input to `builder.add_group()` is the original
group's name appended with a distinct string (`UG` standing for "Unique
Group"), so we can easily retrieve the original group's name.

So, if we had the following control program:
```
control {
  seq { a; a; a; }
}
```
the resulting control program after running `uniquefy-enables` would be:
```
control {
  seq { aUG; aUG0; aUG1; }
}
```

### Par thread assignment

After creating unique groups for each enable, the pass can also assign
thread ids for each enable in a component's control. Anything sequential
to the "main" thread will be assigned 0, and any new `par` arm (even in
distinct `par`s) will yield a new value.
This assignment can be outputted to a file (or stdout/stderr) by using
the `par-thread-json` option.

For example, if the `main` component's control program of `foo.futil`
looked like this:
```
control {
  seq {
     a;
     par { b; c; }
     d;
  }
}
```

then running `cargo run -- foo.futil -p uniquefy-enables -x
uniquefy-enables:par-thread-json=foo.json` will yield the following
`foo.json`:
```
{
  "main": {
    "aUG": 0,
    "bUG": 1,
    "cUG": 2,
    "dUG": 0
  }
}
```

### Path descriptors

After creating unique groups for each enable, the pass can also compute
path descriptors (as described in #2273 ) for each enable and each
`par`. This information can be outputted to a file (or stdout/stderr) by
using the `path-descriptor-json` option. The outputted JSON lists for
each component the descriptors of each enable, and the list of
descriptors that are the root of `par`s.

For example, consider `language-tutorial-iterate.futil`, which has the
following control program in `main`:
```
  control {
    seq {
      init;
      while lt.out with cond {
        par {
          seq {
            read;
            upd;
            write;
          }
          incr;
        }
      }
    }
  }
```
When we run `cargo run --
examples/tutorial/language-tutorial-iterate.futil -p uniquefy-enables -x
uniquefy-enables:path-descriptor-json=language-tutorial-iterate.json`,
we get the following `language-tutorial-iterate.json`:
```
{
  "main": {
    "enables": {
      "incrUG": "main.-1-b-1",
      "initUG": "main.-0",
      "readUG": "main.-1-b-0-0",
      "updUG": "main.-1-b-0-1",
      "writeUG": "main.-1-b-0-2"
    },
    "pars": [
      "main.-1-b-"
    ]
  }
}
```