[Cider] Memory Entangling implementation (#2512)
This PR lands the memory entangling feature that we need for some of the
more interesting memory examples, until multiple go/done interfaces are
given proper semantics.

The feature can be used via the `--entangle` flag which should be
provided with the memory definitions to entangle. If no component name
is provided, then it assumed the named memories are in the entrypoint
component. This feature currently must be used in conjunction with data
race detection. Errors on the entangled memory will refer to it by the
name of the "canonical" memory within the group which is the one that is
defined earliest in the program. If initialization values are provided,
they must be given for the canonical memory in a group and will
otherwise be ignored.

If a single thread writes to the same slot from multiple interfaces
within a single cycle, the write to the latest defined memory will win
(if there are multiple threads, the program will crash with a data
race). It is worth noting that this win ordering only happens because
primitive cycle level evaluation occurs in order of definition. Our
examples should probably not rely on this behavior.

This PR also includes some minor changes to color printing handling in
Cider. The default behavior is now to always have color on but it can be
forced off via `--force-color off` since the auto inference will usually
break when used through `fud2` because everything ends up written to
files.