[misc] bump wasm bindgen version (#2513)
We've been getting a "future incompatibility" warning from this package
for a while but it turns out this would be a blocker if we ever wanted
to switch to rust 1.88 (the one with let chains). I opted to bump it to
the most recent version, which it turns out requests the use of
`serde_wasm_bindgen` as the `to_serde` method is now deprecated. The
path to resolution seems to be using `serde_wasm_bindgen::from_value` in
place of `to_serde`. I had to throw in a clone but given that the
jsvalues are basically just a newtype this should be a trivial cost.

Hopefully this just works? I don't actually know how to test the
playground, so I'm hoping one of y'all has insight.