Calyx formatter (#2519)
This PR introduces a compiler CLI command to format Calyx code by
parsing it, building the IR, and pretty-printing it. It can be invoked
with `cargo run -- format <file>.futil`.

Whether this PR is merged or left as a little side project for now is
definitely up for discussion. Even though it doesn't alter any existing
behavior and the additions are minimal, there are some huge limitations
that arguably make the tool largely unusable.

# Limitations
**The formatter removes all comments.** This is a huge limitation
stemming from the fact that comments are ignored by the `pest` parser,
and the potential solutions are arguably bad. Here are some proposals;
looking forward to more discussion:
- **Add comments to every AST node in the `pest` grammar.** This seems
tedious, and pollutes an otherwise clean grammar with hundreds of nodes
that are only used for one, relatively niche purpose, ultimately making
it harder and error-prone to maintain.
- **Modify the grammar programmatically to include comments.** This
would be ideal, but I haven't found a straightforward way to do it. I'm
not sure `pest` exposes tools to do so.
- Something else?

# Implementation notes
**Why not use `tree-sitter` and the LSP grammar?**
The existing `tree-sitter` grammar written for the LSP is incomplete and
not equivalent to the `pest` grammar. This is not an accidental problem
but rather an intrinsic one with maintaining two parallel versions of
the grammar. It seems unreasonably burdensome to maintain formal
equivalence between the two, especially given that the parsers
themselves are quite different (PEG vs. LR). In addition, the LSP
infrastructure is brittle and largely unmaintained.

**Why build the IR at all?**
There is an existing formatting infrastructure in `printer.rs` that
relies on the IR. The three options were: a) create a separate
formatting infrastructure for the AST; b) refactor the existing
infrastructure to be compatible with the AST; and c) build the IR and
then pass it to the existing infrastructure. Option c) seemed to be the
most reasonable with minimal changes, but it is undeniably a little
strange.

---------

Co-authored-by: Pedro Pontes García <pp457@havarti.cs.cornell.edu>