[Profiler] Preserving @pos attributes through optimizations, showing control groups + source locations, using Perfetto protobuf library for timeline view (#2522)
This PR adds new features to the profiler:
- Adding control groups to the timeline view
- Tracking down source locations of control groups' original control
statements
- Using Perfetto's protobuf format instead of the Google Trace event
format for the timeline view
- Also fixed a bug where all control groups were added to the stack
_every cycle_ instead of just the cycles they were active in

Here's an example of the new timeline view on Perfetto on
`tests/correctness/pipelined-mac.futil`:
<img width="991" height="574" alt="image"
src="https://github.com/user-attachments/assets/e98c5df0-74da-4d0e-a821-f210509396c5"
/>

## Tracking source locations of control groups

To obtain source locations of control groups, we need to (1) know the
source locations of control _nodes_ in the _original Calyx program_, and
(2) associate those source locations with the control _group_ made by
TDCC _after optimizations_.

For (1) I'm using the `metadata-table-generation` pass to add `@pos`
annotations to each control node. I've also updated the
`fileinfo_emitter` tool to process Calyx-file positions differently than
Calyx-py-file positions (previously the only functionality of the tool)
by using filename extensions, which is quite hacky but works for now.
The tool now outputs JSON that maps between the Calyx `@pos` ID, the
line number of the original Calyx control node, and the "type" of
control node that the position maps onto (`par`/`seq`/`while`) to the
file provided with the `-c` option.

ex) For `pipelined-mac.futil` we get the following from the
`fileinfo_emitter` tool:
```
{
  "pipelined_mac": [
    {
      "pos_num": 9,
      "linenum": 74,
      "ctrl_node": "seq"
    }
...
```

For (2) I've updated optimization passes (`static-promotion`,
`static-inline`, `simplify-with-control`) that change control nodes to
propagate original `@pos` attributes to the new control node. I've also
updated TDCC to obtain the `@pos` attributes of the control node that
corresponds to the created control group (the topmost control node of
each `Schedule`), and output it in the `dump-fsm-json` option, which is
used by the profiler.

The profiler script then takes both information from
`metadata-table-generation`/the fileinfo tool and TDCC using the `@pos`
IDs to connect the created control group to the line number and control
node type, which is shown in the timeline view. (The profiler script
currently doesn't depend on information from the fileinfo tool)

## Next steps

Currently the profiler contains a bug where the control groups,
especially TDCC groups that manage seqs/whiles/ifs are not placed on the
stack in order (parent-child relationships of control groups are not
respected). To fix this, I will be using control descriptors obtained
from the `uniquefy-enable` pass (and relying on the position tags given
by the `metadata-table-generation` pass).

At some point in the future, I think I should replace the hacky parts of
my code with principled solutions:
- Using the extension of the filename to categorize positions by ADL
- Using a single tool for both Calyx-py and Calyx sourceinfo processing
(maybe this is ok?)
- **A larger problem is that I'm starting to use `@pos` source info as
both a means to (1) source map and (2) uniquely identify control nodes
in Calyx.** The `@pos` source info should technically be reserved for
(1).

---------

Co-authored-by: Griffin Berlstein <griffin@berlste.in>